create validate id inside token in routes card agent

+++ src/modules/Agents/UseCases/DeleteCardAgent/DeleteCardAgentUseCase.ts
@@ -8,8 +8,9 @@ class DeleteCardAgentUseCase{
   constructor(cardAgentRepository:ICardAgentRepository){
     this.cardAgentRepository = cardAgentRepository
   }
-  async execute(id_agent:string):Promise<CardAgent>{ 
+  async execute({id_agent,id_agent_token}):Promise<CardAgent>{ 
     if(!id_agent) throw new AppError("Value of agent is undefined.")
+    if(id_agent !== id_agent_token) throw new AppError("Token sen not to own agent authenticate ")
     const findCardAgent = await this.cardAgentRepository.listByid(id_agent)
     if(!findCardAgent) throw new AppError("Card Agent not found.")
     const deletedCardAgent = await this.cardAgentRepository.delete(id_agent)
